ref: TokenRequestStrategy 에서는 더이상 클라이언트 정보를 가지지 않는다.
