+++ src/main/java/nextstep/app/application/google/GoogleTokenRequestStrategy.java
@@ -2,7 +2,8 @@ package nextstep.app.application.google;
 
 import nextstep.app.application.google.dto.GoogleTokenRequest;
 import nextstep.app.application.google.dto.GoogleTokenResponse;
-import nextstep.security.authentication.OAuth2TokenRequestStrategy;
+import nextstep.security.authentication.oauth.OAuth2AuthorizationRequest;
+import nextstep.security.authentication.oauth.OAuth2TokenRequestStrategy;
 import nextstep.security.authentication.TokenRequest;
 import nextstep.security.authentication.TokenResponse;
 import org.springframework.beans.factory.annotation.Value;
@@ -11,12 +12,6 @@ import org.springframework.stereotype.Service;
 @Service
 public class GoogleTokenRequestStrategy implements OAuth2TokenRequestStrategy {
 
-    @Value("${oauth2.google.client-id}")
-    private String clientId;
-    @Value("${oauth2.google.client-secret}")
-    private String clientSecret;
-    @Value("${oauth2.google.token.redirect-uri}")
-    private String redirectUri;
     @Value("${oauth2.google.token.request-uri}")
     private String requestUri;
 
@@ -26,8 +21,8 @@ public class GoogleTokenRequestStrategy implements OAuth2TokenRequestStrategy {
     }
 
     @Override
-    public TokenRequest requestToken(String code) {
-        return GoogleTokenRequest.of(code, clientId, clientSecret, redirectUri);
+    public TokenRequest requestToken(OAuth2AuthorizationRequest request, String code) {
+        return GoogleTokenRequest.of(code, request.clientId(), request.clientSecret(), request.redirectUri());
     }
 
     @Override
