--- src/main/java/nextstep/security/authentication/OAuth2TokenRequestStrategy.java
@@ -1,8 +1,0 @@
-package nextstep.security.authentication;
-
-public interface OAuth2TokenRequestStrategy {
-    String getOAuth2Type();
-    TokenRequest requestToken(String code);
-    String getRequestUri();
-    Class<? extends TokenResponse> getResponseClass();
-}
