--- src/main/java/nextstep/security/authentication/OAuth2TokenRequester.java
@@ -1,5 +1,0 @@
-package nextstep.security.authentication;
-
-public interface OAuth2TokenRequester {
-    TokenResponse request(String oAuth2Type, String code);
-}
