+++ src/main/java/nextstep/security/authentication/oauth/OAuth2TokenRequester.java
@@ -1,0 +1,7 @@
+package nextstep.security.authentication.oauth;
+
+import nextstep.security.authentication.TokenResponse;
+
+public interface OAuth2TokenRequester {
+    TokenResponse request(OAuth2AuthorizationRequest authorizationRequest, String code);
+}
